internal/gradle/plugin tidy-up  - moved PluginDiscoveryTask to public api  - introduced internal/gradle/plugin/tasks and moved classes
