+++ src/main/groovy/org/shipkit/gradle/plugin/PluginDiscoveryTask.java
@@ -1,7 +1,8 @@
-package org.shipkit.internal.gradle.plugin;
+package org.shipkit.gradle.plugin;
 
 import org.gradle.api.DefaultTask;
 import org.gradle.api.tasks.TaskAction;
+import org.shipkit.internal.gradle.plugin.tasks.PluginDiscovery;
 
 public class PluginDiscoveryTask extends DefaultTask {
